Add page titles, issue #39
