+++ logicle/app/profile/password/page.tsx
@@ -1,4 +1,9 @@
 import UpdatePassword from '@/components/app/UpdatePassword'
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Change Password',
+};
 
 const Password = () => {
   return <UpdatePassword />
